src/llm-tongyi/requirements.txt
brotli==1.1.0
dashscope==1.24.0
docx2txt==0.9
faiss-cpu==1.11.0.post1
greenlet==3.2.3
h2==4.2.0
jaraco.collections==5.1.0
langchain-anthropic==0.3.17
langchain-community==0.3.27
langchain-huggingface==0.3.1
langgraph==0.5.4
langgraph-checkpoint-sqlite==2.0.10
pip-chill==1.0.3
platformdirs==4.2.2
pysocks==1.7.1
sentence-transformers==5.0.0
tavily-python==0.7.10
tomli==2.0.1
